fix: resolve 28 golangci-lint issues (errcheck, staticcheck, ineffassign, unconvert, unused)
errcheck (11):
- cmd/ethercat: check errors from flag.Parse() in all command handlers
- eni/eni.go: capture Close() error via named return + defer
- internal/link/udp: check Close(), SetReadDeadline(), WriteTo() errors

ineffassign (1):
- internal/link/udp: remove ineffectual 'err = nil' before break

staticcheck (6):
- ecmd: rename NoFrame→ErrNoFrame, NoOverlay→ErrNoOverlay (ST1012)
- ecmd/mux: remove embedded field selector (QF1008)
- ecfr/ecfr_test: omit type int from declaration (QF1011)
- internal/sim/sim_test: omit nil check in len() (S1009)

unconvert (3):
- ecfr/datagram: remove uint16() cast on Interrupt
- internal/sim/l2slave: remove uint8() and uint16() casts

unused (7):
- ecmd/ecmd_test: remove unused mockCommander type+methods, makeLenDgram
- internal/link/udp/udp_test: remove unused findValidInterface
- Keep mockCmd type (used by mockCommanderWithData)

All tests pass, go vet clean

+++ ecfr/datagram.go
@@ -244,7 +244,7 @@ func (dh *DatagramHeader) Commit() ([]byte, error) {
 	}
 	lo := uint64(dh.Command) | uint64(dh.Index)<<8 | uint64(dh.Addr32)<<16 | uint64(dh.LenWord)<<48
 	*(*uint64)(unsafe.Pointer(&dh.buffer[0])) = lo
-	*(*uint16)(unsafe.Pointer(&dh.buffer[8])) = uint16(dh.Interrupt)
+	*(*uint16)(unsafe.Pointer(&dh.buffer[8])) = dh.Interrupt
 	return dh.buffer[:datagramHeaderLength], nil
 }
